ARM: ftrace: Ensure code modifications are synchronised across all cpus
When the generic ftrace implementation modifies code for trace-points it

uses stop_machine() to call ftrace_modify_all_code() on one CPU. This

ultimately calls the ARM specific function ftrace_modify_code() which

updates the instruction and then does flush_icache_range(). As this

cache flushing only operates on the local CPU then other cores may end

up execute the old instruction if it's still in their icaches.

This may or may not cause problems for the use of ftrace on kernels

compiled for ARM instructions. However, Thumb2 instructions can straddle

two cache lines so its possible for half the old instruction to be in

the cache and half the new one, leading the the CPU executing garbage.

This patch fixes this situation by providing and arch specific

implementation of arch_ftrace_update_code() which ensures that after one

core has modified all the code, the other cores invalidate their icaches

before continuing.

Signed-off-by: Jon Medhurst <tixy@linaro.org>